Create build.sh
Script to build a docker image from scratch. This script pulls the latest version of the inline scanner, makes it executable, then builds the image. The Image Name variable is currently set to Julien's DockerHub repo. This will need to be change to point to the Lacework repo/image-name chosen.